requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.6.3
blinker==1.8.2
cachetools==4.2.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
gevent==23.9.1
google-api-core==2.17.1
google-api-python-client==2.136.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
gspread==6.1.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.35.7
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyngrok==7.1.6
pyparsing==3.1.2
python-telegram-bot==13.15
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
tornado==6.1
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
